requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.8.0
click==8.1.7
diff-match-patch==20230430
Django==5.0
django-bootstrap4==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
